+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## v1.6.0
 ### Added
  - `pubSendNotification(notification: any): Promise<void>`
+ - `subSendNotification(callback: (message: any) => void): Promise<void>`
 
 ## v1.5.1 (2020-03-27)
 ### Fixed